requirements.txt
pandas
numpy
streamlit
google-generativeai
scikit-learn
pypdf
python-dotenv
ollama
